requirements.txt
locustio==0.8.1
thriftpy==0.3.8